Move embed type converter
